Move the lockstep_api class into its own folder
